% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{ribo_raw_human_cap_995}
\alias{ribo_raw_human_cap_995}
\title{Ribo count matrix}
\format{
A numeric matrix with G genes and S samples
}
\source{
Yue's Mozart Folder. Assuming that this follows Jonathan's counting
}
\usage{
ribo_raw_human_cap_995
}
\description{
Genes by samples count matrix.
}
\keyword{datasets}
