% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data-doc.R
\docType{data}
\name{rnaseq_raw_human_cap_995}
\alias{rnaseq_raw_human_cap_995}
\title{RNA count matrix}
\format{
A numeric matrix
}
\source{
Yue's Mozart Folder. Assuming that this follows Jonathan's counting
}
\usage{
rnaseq_raw_human_cap_995
}
\description{
RNA count matrix
}
\keyword{datasets}
